requirements.txt
torch
torchvision
numpy
pillow
flask
flask-cors
scikit-learn
opencv-python
